Figma2VectorDB/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
